unihaven-extension/requirements.txt
Django==4.2
djangorestframework==3.14.0
Pillow==10.0.0
python-decouple==3.5
requests==2.31.0
django-cors-headers==3.14.0
django-environ==0.8.1
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
gunicorn==21.2.0